move ExpressionEditor to common
